enabled sequential double buffer
