+++ Cargo.toml
@@ -19,7 +19,7 @@ opt-level = 3
 [dependencies]
 hashbrown = "0.3"
 priority-queue = "0.5.4"
-criterion =  { path = "criterion.rs"}
+criterion =  "0.3.3"
 rand = "0.6.5"
 lazy_static = "1.3.0"
 piston_window = "0.105.0"
